update: mapping earth texture

+++ inc/common.hpp
@@ -13,7 +13,7 @@
 
 namespace gl_learn {
 using Vertex      = std::array<GLfloat, 3>;
-using UV          = std::array<GLuint, 2>;
+using UV          = std::array<GLfloat, 2>;
 
 constexpr std::uint32_t WIDTH  = 800;
 constexpr std::uint32_t HEIGHT = 600;
